+++ src/services/disc.service.js
@@ -6,6 +6,7 @@ import { FileSystemUtils } from "../utils/filesystem.js";
 import { VALIDATION_CONSTANTS, MEDIA_TYPES } from "../constants/index.js";
 import { DriveService } from "./drive.service.js";
 import { MakeMKVMessages } from "../utils/makemkv-messages.js";
+import { createDateEnvironment } from "../utils/process.js";
 
 /**
  * Service for handling disc detection and information gathering
@@ -93,7 +94,11 @@ export class DiscService {
 
       const command = `${makeMKVExecutable} -r info disc:index`;
 
-      exec(command, (err, stdout, stderr) => {
+      // Create environment with fake date if configured
+      const fakeDate = AppConfig.makeMKVFakeDate;
+      const env = { ...process.env, ...createDateEnvironment(fakeDate) };
+
+      exec(command, { env }, (err, stdout, stderr) => {
         // Check for critical MakeMKV messages first
         const isFirstCall = DiscService.isFirstMakeMKVCall;
         const shouldContinue = MakeMKVMessages.checkOutput(
@@ -298,7 +303,11 @@ export class DiscService {
 
       const command = `${makeMKVExecutable} -r info disc:${driveInfo.driveNumber}`;
 
-      exec(command, (err, stdout, stderr) => {
+      // Create environment with fake date if configured
+      const fakeDate = AppConfig.makeMKVFakeDate;
+      const env = { ...process.env, ...createDateEnvironment(fakeDate) };
+
+      exec(command, { env }, (err, stdout, stderr) => {
         // Check for critical MakeMKV messages (not first call, so only check for errors)
         const shouldContinue = MakeMKVMessages.checkOutput(
           stdout + (stderr || ""),